move brico to retrieval
