+++ mteb/tasks/Reranking/eng/__init__.py
@@ -1,11 +1,16 @@
 from __future__ import annotations
 
+from mteb.tasks.Retrieval.eng.BIRCOArguAnaReranking import BIRCOArguAnaReranking
+from mteb.tasks.Retrieval.eng.BIRCOClinicalTrialReranking import (
+    BIRCOClinicalTrialReranking,
+)
+from mteb.tasks.Retrieval.eng.BIRCODorisMaeReranking import BIRCODorisMaeReranking
+from mteb.tasks.Retrieval.eng.BIRCORelicReranking import BIRCORelicReranking
+from mteb.tasks.Retrieval.eng.BIRCOWhatsThatBookReranking import (
+    BIRCOWhatsThatBookReranking,
+)
+
 from .AskUbuntuDupQuestions import AskUbuntuDupQuestions
-from .BIRCOArguAnaReranking import BIRCOArguAnaReranking
-from .BIRCOClinicalTrialReranking import BIRCOClinicalTrialReranking
-from .BIRCODorisMaeReranking import BIRCODorisMaeReranking
-from .BIRCORelicReranking import BIRCORelicReranking
-from .BIRCOWhatsThatBookReranking import BIRCOWhatsThatBookReranking
 from .BuiltBenchReranking import BuiltBenchReranking
 from .MindSmallReranking import MindSmallReranking
 from .NevIR import NevIR
